Assistant checkpoint: Added time estimate message on get-started page
Assistant generated file changes:
- client/src/pages/get-started.tsx: Add informational message about the 5-minute process

---

User prompt:

on the second page. Pls inform the user that it will take about 5 minutes to collect all the information since we are trying to guide them in the best possible way. if they know what they want then they can tell that directly to the AI and it will jump ahead.